Removed NGUI folder and made some more changes.

+++ Assets/Parafall Assets/Scripts/ParafallObjectPool.cs
@@ -21,7 +21,7 @@ public class ParafallObjectPool : MonoBehaviour {
 
 	public List<ParaPacket> listOfPackets;
 
-	private int objectsSpawnedFromPool = 1;
+	private long objectsSpawnedFromPool = 1;
 
 	private static ParafallObjectPool instance;
